+++ dmutils/forms/errors.py
@@ -83,3 +83,74 @@ def govuk_errors(errors: typing.Mapping[str, ErrorMapping]) -> typing_OrderedDic
     return OrderedDict(
         (key, {**error, **govuk_error(error)}) for key, error in errors.items()
     )
+
+
+def get_errors_from_wtform(form):
+    """Converts Flask-WTForm errors into formats suitable for use in Digital Marketplace templates.
+
+    Returns a dictionary with the following keys, supporting three types of error display:
+
+    1) digitalmarketplace-frontend-toolkit template toolkit/forms/validation.html:
+      `input_name`, `question`, `message`
+
+    2) govuk-frontend macro govukErrorSummary:
+      `text`, `href`
+
+    3) govuk-frontend errorMessage:
+      `errorMessage`
+
+    This allows us to treat errors from both content-loader forms and wtforms the same way inside templates,
+    and provides backwards-compatible support for the migration to using govuk-frontend components.
+
+    Example usage:
+
+        # app.py
+        class Form(FlaskForm):
+            input = DMTextInput()
+
+        @flask.route("/")
+        def view():
+            form = Form()
+            errors = get_errors_from_wtform()
+            return render(
+                "template.html",
+                errors=errors,
+            )
+
+        # template.html  (govuk-frontend)
+        {{ govukErrorSummary({
+            "errorList": errors.values()
+        }) }}
+
+        {{ govukTextInput({
+            "errorMessage": errors.input.errorMessage,
+        }) }}
+
+        # template.html (DM frontend toolkit uses 'errors' by default)
+        {% include 'toolkit/forms/validation.html' %}
+
+    :param form: A Flask-WTForm
+    :return: A dict with error information in a form suitable for Digital Marketplace templates
+    """
+    return OrderedDict(
+        # TODO: remove legacy code (items for frontend toolkit validation banners, 'input-' prefix)
+        (
+            key,
+            {
+                # parameters for digitalmarketplace-frontend-toolkit template toolkit/forms/validation.html
+                "input_name": key, "question": form[key].label.text, "message": form[key].errors[0],
+
+                # parameters for govuk-frontend macro govukErrorSummary
+                "text": form[key].errors[0], "href": f"#input-{key}",
+
+                # parameters for govuk-frontend errorMessage parameter
+                "errorMessage": (
+                    {"text": form[key].errors[0]}
+                    if form[key].errors[0]
+                    else {}
+                )
+            }
+        )
+        for key in
+        form.errors.keys()
+    )
